gptonline/requirements.txt
openai==0.27.4
requests==2.31.0
python-dotenv==1.0.0
tavily-python
beautifulsoup4==4.12.2
streamlit==1.24.1
